requirements.txt
greenlet==3.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
